sprite적용, 쿨쇼, 스테이지

+++ yesterday-shooting/Assets/02.Scripts/Item/ItemSkill.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public abstract class ItemSkill : MonoBehaviour
 {
-    public int maxCool;
+    public float maxCool;
     public abstract bool Skill();
 }
